// src/app/model/employee/family.ts
export interface family {
  id?: number;

  relationShip?: string; // Quan hệ
  name?: string; // Họ tên"
  birthday?: string; // Ngày sinh"
  gender?: number; // Giới tính"
  job?: string; // Nghề nghiệp"
  workUnit?: string; // Nơi công tác/làm việc"
  permanentAddress?: string; // Hộ khẩu thường trú"
  currentAddress?: string; // Nơi ở hiện tại"
  cardId?: string; // Số CMT/Thẻ căn cước"
  cardDate?: string; // Ngày cấp Số CMT/Thẻ căn cước"
  cardAddress?: string; // Nơi cấp Số CMT/Thẻ căn cước"
  passportNumber?: string; // Số sổ hộ chiếu"
  passportDate?: string; // Ngày cấp hộ chiếu"
  passportExpiredDate?: string; // Ngày hết hạn hộ chiếu"
  homePhone?: string; // Điện thoại nhà riêng"
  mobilePhone?: string; // Điện thoại di động"
  email?: string; // Email BVB"
  emailPrivate?: string; // Email cá nhân"
  description?: string; // "

  employee?: {};
}